refactor questionnaire config
