+++ src/config/questionnaire_config.js
@@ -1,8 +1,42 @@
-import { capitalizeFirstLetterSafe, isEmptyArray, isNil } from "@util";
+import { isEmptyArray } from "@util";
 import { getResourcesByResourceType, linkIdEquals } from "@util/fhirUtil";
-import CHART_CONFIG, { SUBSTANCE_USE_LINE_PROPS } from "./chart_config";
-import { PHQ9_SI_QUESTION_LINK_ID, PHQ9_SI_ANSWER_SCORE_MAPPINGS, PHQ9_ADMIN_NOTE } from "@/consts";
+import CIRG_ADL_IADL from "./questionnaire_configs/CIRG_ADL_IADL";
+import CIRG_BEHAV5 from "./questionnaire_configs/CIRG_BEHAV5";
+import CIRG_C_IDAS from "./questionnaire_configs/CIRG_C_IDAS";
+import CIRG_CNICS_ARV, { CIRG_CNICS_ARV_MISSED_DOSE, CIRG_CNICS_ARV_SRS, CIRG_CNICS_ARV_VAS } from "./questionnaire_configs/CIRG_CNICS_ARV";
+import CIRG_CNICS_ASSIST from "./questionnaire_configs/CIRG_CNICS_ASSIST";
+import CIRG_CNICS_ASSIS_OD from "./questionnaire_configs/CIRG_CNICS_ASSIS_OD";
+import CIRG_CNICS_ASSIST_POLYSUB from "./questionnaire_configs/CIRG_CNICS_ASSIST_POLYSUB";
+import CIRG_CNICS_AUDIT from "./questionnaire_configs/CIRG_CNICS_AUDIT";
+import CIRG_CNICS_EUROQOL, {
+  CIRG_CNICS_EUROQOL_ANXIETY_DEPRESSION,
+  CIRG_CNICS_EUROQOL_EUROQOL_5,
+  CIRG_CNICS_EUROQOL_PAIN_DISCOMFORT,
+  CIRG_CNICS_EUROQOL_SELF_CARE,
+  CIRG_CNICS_EUROQOL_USUAL_ACTIVITIES,
+} from "./questionnaire_configs/CIRG_CNICS_EUROQOL";
+import CIRG_CNICS_EXCHANGE_SEX from "./questionnaire_configs/CIRG_CNICS_EXCHANGE_SEX";
+import CIRG_CNICS_FINANCIAL from "./questionnaire_configs/CIRG_CNICS_FINANCIAL";
+import CIRG_CNICS_FOOD from "./questionnaire_configs/CIRG_CNICS_FOOD";
+import CIRG_CNICS_FROP_COM from "./questionnaire_configs/CIRG_CNICS_FROP_COM";
+import CIRG_CNICS_HIV_STIGMA from "./questionnaire_configs/CIRG_CNICS_HIV_STIGMA";
+import CIRG_CNICS_HOUSING from "./questionnaire_configs/CIRG_CNICS_HOUSING";
+import CIRG_CNICS_IPV4 from "./questionnaire_configs/CIRG_CNICS_IPV4";
+import CIRG_CNICS_MAPSS_SF from "./questionnaire_configs/CIRG_CNICS_MAPSS_SF";
+import CIRG_CNICS_MINI from "./questionnaire_configs/CIRG_CNICS_MINI";
+import CIRG_CNICS_SEXUAL_RISK, { CIRG_SEXUAL_PARTNER_CONTEXT, CIRG_SEXUAL_PARTNERS, CIRG_STI, CIRG_UNPROTECTED_ANAL_SEX, CIRG_UNPROTECTED_ORAL_SEX, CIRG_UNPROTECTED_VAGINAL_SEX } from "./questionnaire_configs/CIRG_CNICS_SEXUAL_RISK";
+import CIRG_CNICS_SMOKING from "./questionnaire_configs/CIRG_CNICS_SMOKING";
+import CIRG_CNICS_SYMPTOMS from "./questionnaire_configs/CIRG_CNICS_SYMPTOMS";
+import CIRG_CP_ECOG from "./questionnaire_configs/CIRG_CP_ECOG";
+import CIRG_ECOG12 from "./questionnaire_configs/CIRG_ECOG12";
+import CIRG_GAD7 from "./questionnaire_configs/CIRG_GAD7";
+import CIRG_GDS from "./questionnaire_configs/CIRG_GDS";
+import CIRG_MINICOG from "./questionnaire_configs/CIRG_MINICOG";
+import CIRG_PC_PTSD_5 from "./questionnaire_configs/CIRG_PC_PTSD_5";
+import CIRG_PHQ9, { CIRG_SI } from "./questionnaire_configs/CIRG_PHQ9";
+import CIRG_SLUMS from "./questionnaire_configs/CIRG_SLUMS";
 import QuestionnaireScoringBuilder from "@/models/resultBuilders/QuestionnaireScoringBuilder";
+
 
 function normalizeInstrumentConfigKeys(config) {
   if (!config || typeof config !== "object") return config;
@@ -249,1489 +283,70 @@ function bootstrapInstrumentConfigMap(map) {
  * @param {object[]} [yLineFields] additional line fields to show in chart
  */
 const questionnaireConfigsRaw = {
-  "CIRG-ADL-IADL": {
-    key: "CIRG-ADL-IADL",
-    questionnaireId: "CIRG-ADL-IADL",
-    questionnaireName: "adl-iadl",
-    instrumentName: "ADL-IADL",
-    title: "ADL IADL",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/adl-iadl",
-    scoringQuestionId: "adl-iadls-total-score",
-    maximumScore: 45,
-    questionLinkIds: [
-      "/46595-5",
-      "/46597-1",
-      "/46599-7",
-      "/57243-8",
-      "/57244-6",
-      "/57246-1",
-      "/57247-9",
-      "/57248-7",
-      "/57249-5",
-      "/46569-0",
-    ],
-    questionnaireMatchMode: "fuzzy",
-    chartParams: { ...CHART_CONFIG.default, xLabel: "" },
-  },
-
-  "CIRG-BEHAV5": {
-    key: "CIRG-BEHAV5",
-    questionnaireId: "CIRG-BEHAV5",
-    questionnaireName: "behav5",
-    instrumentName: "BEHAV-5",
-    title: "BEHAV-5",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/behav5",
-    scoringQuestionId: "behav-8",
-    maximumScore: 6,
-    questionLinkIds: ["behav-1", "behav-2", "behav-3", "behav-4", "behav-5", "behav-6"],
-    questionnaireMatchMode: "fuzzy",
-    chartParams: { ...CHART_CONFIG.default, minimumYValue: 0, maximumYValue: 6, xLabel: "" },
-  },
-
-  "CIRG-C-IDAS": {
-    key: "CIRG-C-IDAS",
-    questionnaireId: "CIRG-C-IDAS",
-    questionnaireName: "c-idas",
-    instrumentName: "C-IDAS",
-    title: "C-IDAS",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/c-idas",
-    scoringQuestionId: "c-ids-score",
-    maximumScore: 36,
-    questionnaireMatchMode: "fuzzy",
-    chartParams: { ...CHART_CONFIG.default, minimumYValue: 0, maximumYValue: 36, xLabel: "" },
-  },
-
-  "CIRG-CNICS-ARV": {
-    key: "CIRG-CNICS-ARV",
-    questionnaireId: "CIRG-CNICS-ARV",
-    title: "Antiretroviral (ARV) adherence",
-    instrumentName: "CIRG-CNICS-ARV",
-  },
-
-  "CIRG-CNICS-ASSIST": {
-    key: "CIRG-CNICS-ASSIST",
-    instrumentName: "Substances Reported",
-    questionnaireMatchMode: "strict",
-    title: "Substances Reported",
-    subtitle: "Past 3 months",
-    disableHeaderRowSubtitle: true,
-    minimumScore: 0,
-    maximumScore: 4,
-    meaningQuestionId: "ASSIST-3mo-score",
-    meaningRowLabel: "Substances Reported (Past 3 months)",
-    nullScoreAllowed: true,
-    questionRowLabel: "Detailed Responses",
-    fallbackScoreMap: {
-      "assist-10-0": 0,
-      "assist-10-1": 1,
-      "assist-10-2": 2,
-      "assist-10-3": 3,
-      "assist-10-4": 4,
-      "assist-11-0": 0,
-      "assist-11-1": 1,
-      "assist-11-2": 2,
-      "assist-11-3": 3,
-      "assist-11-4": 4,
-      "assist-12-0": 0,
-      "assist-12-1": 1,
-      "assist-12-2": 2,
-      "assist-12-3": 3,
-      "assist-12-4": 4,
-      "assist-13-0": 0,
-      "assist-13-1": 1,
-      "assist-13-2": 2,
-      "assist-13-3": 3,
-      "assist-13-4": 4,
-      "assist-14-0": 0,
-      "assist-14-1": 1,
-      "assist-14-2": 2,
-      "assist-14-3": 3,
-      "assist-14-4": 4,
-      "assist-15-0": 0,
-      "assist-15-1": 1,
-      "assist-15-2": 2,
-      "assist-15-3": 3,
-      "assist-15-4": 4,
-      "assist-16-0": 0,
-      "assist-16-1": 1,
-      "assist-16-2": 2,
-      "assist-16-3": 3,
-      "assist-16-4": 4,
-      "assist-17-0": 0,
-      "assist-17-1": 1,
-      "assist-17-2": 2,
-      "assist-17-3": 3,
-      "assist-17-4": 4,
-      "assist-18-0": 0,
-      "assist-18-1": 1,
-      "assist-18-2": 2,
-      "assist-18-3": 3,
-      "assist-18-4": 4,
-      "assist-19-0": 0,
-      "assist-19-1": 1,
-      "assist-19-2": 2,
-      "assist-19-3": 3,
-      "assist-19-4": 4,
-    },
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const meaningResponse = responses.find((response) => linkIdEquals(response.id, "ASSIST-3mo-score", "strict"));
-      const meaningAnswer =
-        meaningResponse?.answer != null && meaningResponse.answer !== undefined ? meaningResponse.answer : null;
-      return meaningAnswer?.split(",").join("|");
-    },
-    subScoringQuestions: [
-      {
-        key: "ASSIST-10",
-        linkId: "ASSIST-10",
-      },
-      {
-        key: "ASSIST-11",
-        linkId: "ASSIST-11",
-      },
-      {
-        key: "ASSIST-12",
-        linkId: "ASSIST-12",
-      },
-      {
-        key: "ASSIST-13",
-        linkId: "ASSIST-13",
-      },
-      {
-        key: "ASSIST-14",
-        linkId: "ASSIST-14",
-      },
-      {
-        key: "ASSIST-15",
-        linkId: "ASSIST-15",
-      },
-      {
-        key: "ASSIST-16",
-        linkId: "ASSIST-16",
-      },
-      {
-        key: "ASSIST-17",
-        linkId: "ASSIST-17",
-      },
-      {
-        key: "ASSIST-18",
-        linkId: "ASSIST-18",
-      },
-      {
-        key: "ASSIST-19",
-        linkId: "ASSIST-19",
-      },
-    ],
-    //legendIconType: "circle",
-    yLineFields: [
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-10",
-        label: "Cocaine/crack",
-        color: "#0072B2",
-      },
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-11",
-        label: "Methamphetamine",
-        color: "#D55E00",
-      },
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-12",
-        label: "Heroin",
-        color: "#009E73",
-      },
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-13",
-        label: "Fentanyl",
-        color: "#CC79A7",
-      },
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-14",
-        color: "#56B4E9",
-        label: "Narcotic pain meds",
-      },
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-15",
-        label: "Sedatives",
-        color: "#E69F00",
-      },
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-16",
-        label: "Marijuana",
-        color: "#8884FF",
-      },
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-17",
-        color: "#545E56",
-        label: "Stimulants",
-      },
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-18",
-        color: "#FF5555",
-        label: "Inhalants",
-      },
-      {
-        ...SUBSTANCE_USE_LINE_PROPS,
-        key: "ASSIST-19",
-        color: "#3DB6B1",
-        label: "Psychedelics",
-      },
-    ],
-    displayMeaningNotScore: true,
-    xLabel: "",
-    linkIdMatchMode: "strict",
-    chartParams: {
-      ...CHART_CONFIG.default,
-      title: "Substances Reported",
-      chartHeight: 420,
-      interval: 0,
-      xsChartHeight: 460,
-      minimumYValue: 0,
-      maximumYValue: 4,
-      chartMargin: {
-        top: 52,
-        right: 20,
-        left: 20,
-        bottom: 10,
-      },
-      yLabel: "Frequency",
-      yLabelVisible: true,
-      yTickFormatter: (value) => {
-        const labels = {
-          0: "Never",
-          1: "Once/Twice",
-          2: "Monthly",
-          3: "Weekly",
-          4: "Daily",
-        };
-        return labels[value] || value;
-      },
-      yTicks: [0, 1, 2, 3, 4],
-      showTooltipMeaning: false,
-      isCategoricalY: true,
-      // enableAxisMeaningLabels: true,
-      yLabelProps: {
-        position: "top",
-        angle: 0,
-        dy: -16,
-        fontSize: "10px",
-      },
-      showYTicks: true,
-      connectNulls: true,
-      tooltipValueFormatter: (value) => {
-        if (value == null || value === undefined) return null;
-        if (value === 0) return "Never";
-        if (value === 1) return "Once or twice";
-        if (value === 2) return "Monthly";
-        if (value === 3) return "Weekly";
-        if (value === 4) return "Daily or almost daily";
-        return null;
-      },
-      wrapperClass: "big",
-      dotRadius: 5,
-      activeDotRadius: 6,
-      //  lineType: "stepAfter"
-      // lineType: "stepBefore",
-      lineType: "monotone",
-      jitterSpreadDays: 16,
-      enableLineSwitches: true,
-      noDataText: "Not Answered",
-    },
-  },
-
-  "CIRG-CNICS-ASSIST-OD": {
-    key: "CIRG-CNICS-ASSIST-OD",
-    instrumentName: "CNICS ASSIST Overdose",
-    title: "Overdose",
-    subtitle: "Past 6 months",
-    disableHeaderRowSubtitle: true,
-    questionnaireMatchMode: "strict",
-    highSeverityScoreCutoff: 1,
-    comparisonToAlert: "higher",
-    displayMeaningNotScore: true,
-    scoringQuestionId: "ASSIST-OD-recent",
-    linkIdMatchMode: "strict",
-    fallbackScoreMap: {
-      "assist-od-recent-0": 0,
-      "assist-od-recent-1": 1,
-      "assist-od-recent-2": 2,
-      "assist-od-recent-3": 3,
-      "assist-od-recent-4": 4,
-    },
-    severityBands: [
-      { min: 1, label: "high", meaning: "Yes, overdose" },
-      { min: 0, label: "low", meaning: "No overdose" },
-    ],
-    meaningRowLabel: "Summary (Past 6 months)",
-    skipChart: true,
-  },
-
-  "CIRG-CNICS-ASSIST-Polysub": {
-    key: "CIRG-CNICS-ASSIST-Polysub",
-    instrumentName: "CNICS Polysubstance Use",
-    title: "Concurrent Drug Use",
-    subtitle: "Past 3 months",
-    questionnaireMatchMode: "strict",
-    highSeverityScoreCutoff: 1,
-    displayMeaningNotScore: true,
-    scoringQuestionId: "ASSIST-Polysub",
-    linkIdMatchMode: "strict",
-    fallbackScoreMap: {
-      "assist-polysub-0": 1,
-      "assist-polysub-1": 0,
-    },
-    severityBands: [
-      { min: 1, label: "high", meaning: "Yes, concurrent drug use" },
-      { min: 0, label: "low", meaning: "No, no concurrent drug use" },
-    ],
-    meaningRowLabel: "Summary",
-    skipChart: true,
-  },
-
-  "CIRG-CNICS-AUDIT": {
-    title: "Alcohol Score",
-    subtitle: "Past 12 months",
-    key: "CIRG-CNICS-AUDIT",
-    instrumentName: "CNICS AUDIT (alcohol consumption questions)",
-    minimumScore: 0,
-    maximumScore: 38,
-    scoringQuestionId: "AUDIT-score",
-    subScoringQuestions: [
-      {
-        key: "AUDIT-C-score",
-        linkId: "AUDIT-C-score",
-      },
-      {
-        key: "AUDIT-score",
-        linkId: "AUDIT-score",
-      },
-    ],
-    displayMeaningNotScore: true,
-    linkIdMatchMode: "strict",
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      let arrMeaning = [];
-      const scoreToReportResponse = responses.find((response) =>
-        linkIdEquals(response.id, "AUDIT-qnr-to-report", "strict"),
-      );
-      const scoreToReport = scoreToReportResponse ? scoreToReportResponse.answer : null;
-      const interpretationResponse = responses.find((response) =>
-        linkIdEquals(
-          response.id,
-          scoreToReport === "AUDIT-C" ? "AUDIT-C-score-interpretation" : "AUDIT-score-interpretation",
-          "strict",
-        ),
-      );
-      const interpretation = interpretationResponse ? interpretationResponse.answer : null;
-      const auditCResponse = responses.find((response) => linkIdEquals(response.id, "AUDIT-C-score", "strict"));
-      const auditCScore = auditCResponse ? parseInt(auditCResponse.answer, 10) : null;
-      const auditResponse = responses.find((response) => linkIdEquals(response.id, "AUDIT-score", "strict"));
-      const auditScore = auditResponse ? parseInt(auditResponse.answer, 10) : null;
-      if (scoreToReport === "AUDIT-C" && auditCScore != null) {
-        arrMeaning.push(auditCScore + " (AUDIT-C)");
-      } else if (scoreToReport === "AUDIT" && auditScore != null) {
-        arrMeaning.push(auditScore + " (AUDIT)");
-      }
-      if (!isEmptyArray(arrMeaning) && interpretation) {
-        arrMeaning.push(interpretation);
-      }
-      return arrMeaning.join("|");
-    },
-    chartParams: {
-      ...CHART_CONFIG.default,
-      title: "Alcohol Score",
-      minimumYValue: 0,
-      maximumYValue: 38,
-      xLabel: "",
-      scoringQuestionId: "AUDIT-score",
-      yLineFields: [
-        {
-          key: "AUDIT-C-score",
-          color: "#498C8A",
-          strokeWidth: 1,
-          legendType: "line",
-          strokeOpacity: 0.6,
-          //strokeDasharray: "2 2", // dashed line
-        },
-        {
-          key: "AUDIT-score",
-          color: "#600876",
-          strokeWidth: 1,
-          strokeOpacity: 0.6,
-          legendType: "line",
-        },
-      ],
-      showTooltipMeaning: false,
-    },
-  },
-
-  "CIRG-CNICS-EUROQOL": {
-    key: "CIRG-CNICS-EUROQOL",
-    instrumentName: "HRQOL",
-    title: "EuroQOL Health Related Quality-of-Life questionnaire",
-  },
-
-  "CIRG-CNICS-EUROQOL-ANXIETY-DEPRESSION": {
-    key: "CIRG-CNICS-EUROQOL-ANXIETY-DEPRESSION",
-    instrumentName: "Anxiety/Depression",
-    title: "Anxiety/Depression",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-EUROQOL"], // one or many hosts
-      linkId: "EUROQOL-SCORE-ANXIETY-DEPRESSION", // the single item to keep
-    },
-    meaningQuestionId: "EUROQOL-SCORE-ANXIETY-DEPRESSION",
-    skipChart: true,
-    skipResponses: true,
-    displayMeaningNotScore: true,
-    meaningRowLabel: "EUROQOL: Anxiety/Depression",
-  },
-
-  "CIRG-CNICS-EUROQOL-EUROQOL-5": {
-    key: "CIRG-CNICS-EUROQOL-EUROQOL-5",
-    instrumentName: "CIRG-CNICS-EUROQOL-EUROQOL-5",
-    title: "Overall Health State",
-    subtitle: "(0% - 100%)",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-EUROQOL"], // one or many hosts
-      linkId: "EUROQOL-5", // the single item to keep
-    },
-    meaningQuestionId: "EUROQOL-5",
-    skipChart: true,
-    skipResponses: true,
-    displayMeaningNotScore: true,
-    valueFormatter: (value) => (!isNil(value) ? `${value} ${String(value).includes("%") ? "" : "%"}` : value),
-    meaningRowLabel: "EUROQOL: Overall health state (0% - 100%)",
-  },
-
-  "CIRG-CNICS-EUROQOL-PAIN-DISCOMFORT": {
-    key: "CIRG-CNICS-EUROQOL-PAIN-DISCOMFORT",
-    instrumentName: "Pain/Discomfort",
-    title: "Pain/Discomfort",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-EUROQOL"], // one or many hosts
-      linkId: "EUROQOL-SCORE-PAIN", // the single item to keep
-    },
-    meaningQuestionId: "EUROQOL-SCORE-PAIN",
-    skipChart: true,
-    skipResponses: true,
-    displayMeaningNotScore: true,
-    meaningRowLabel: "EUROQOL: Pain/Discomfort",
-  },
-
-  "CIRG-CNICS-EUROQOL-SELF-CARE": {
-    key: "CIRG-CNICS-EUROQOL-SELF-CARE",
-    instrumentName: "Self Care",
-    title: "Self Care",
-    meaningQuestionId: "EUROQOL-SCORE-SELF-CARE",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-EUROQOL"], // one or many hosts
-      linkId: "EUROQOL-SCORE-SELF-CARE", // the single item to keep
-    },
-    skipChart: true,
-    skipResponses: true,
-    displayMeaningNotScore: true,
-    meaningRowLabel: "EUROQOL: Self Care",
-  },
-
-  "CIRG-CNICS-EUROQOL-USUAL-ACTIVITIES": {
-    key: "CIRG-CNICS-EUROQOL-USUAL-ACTIVITIES",
-    instrumentName: "Usual Activities",
-    title: "Usual Activities",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-EUROQOL"], // one or many hosts
-      linkId: "EUROQOL-SCORE-USUAL-ACTIVITIES", // the single item to keep
-    },
-    meaningQuestionId: "EUROQOL-SCORE-USUAL-ACTIVITIES",
-    skipChart: true,
-    skipResponses: true,
-    displayMeaningNotScore: true,
-    meaningRowLabel: "EUROQOL: Usual Activities",
-  },
-
-  "CIRG-CNICS-EXCHANGE-SEX": {
-    key: "CIRG-CNICS-EXCHANGE-SEX",
-    insturmentName: "Exchange Sex",
-    title: "Exchange Sex",
-    subtitle: "Past 3 months",
-    columns: [
-      {
-        linkId: "EXCHANGE-SEX-SCORE-PAST-3-MONTHS",
-        id: "result",
-      },
-    ],
-    displayMeaningNotScore: true,
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const meaningResponse = responses.find((response) =>
-        linkIdEquals(response.id, "EXCHANGE-SEX-SCORE-PAST-3-MONTHS", "strict"),
-      );
-      const meaningAnswer =
-        meaningResponse?.answer != null && meaningResponse.answer !== undefined ? meaningResponse.answer : null;
-      if (String(meaningAnswer).toLowerCase() === "true") return "Yes";
-      else if (String(meaningAnswer).toLowerCase() === "false") return "No";
-      return meaningAnswer;
-    },
-    linkIdMatchMode: "strict",
-    meaningQuestionId: "EXCHANGE-SEX-SCORE-PAST-3-MONTHS",
-    meaningRowLabel: "Exchange Sex (Past 3 months)",
-    disableHeaderRowSubtitle: true,
-    skipChart: true,
-  },
-
-  "CIRG-CNICS-FINANCIAL": {
-    key: "CIRG-CNICS-FINANCIAL",
-    instrumentName: "CNICS Financial Situation Questionnaire",
-    title: "Financial Situation",
-    questionnaireMatchMode: "fuzzy",
-    highSeverityScoreCutoff: 1,
-    comparisonToAlert: "higher",
-    displayMeaningNotScore: true,
-    linkIdMatchMode: "strict",
-    fallbackScoreMap: {
-      "FINANCIAL-0-0": 1,
-      "FINANCIAL-0-1": 0,
-      "FINANCIAL-0-2": 0,
-      "FINANCIAL-0-3": 0,
-    },
-    alertQuestionId: "FINANCIAL-critical-flag",
-    meaningQuestionId: "FINANCIAL-score-label",
-    skipResponses: true,
-    skipChart: true,
-  },
-
-  "CIRG-CNICS-FOOD": {
-    key: "CIRG-CNICS-FOOD",
-    instrumentName: "CNICS Food Security Questionnaire",
-    title: "Food Security",
-    subtitle: "Past 12 months",
-    questionnaireMatchMode: "fuzzy",
-    displayMeaningNotScore: true,
-    scoringQuestionId: "FOOD-score",
-    alertQuestionId: "FOOD-critical-flag",
-    meaningQuestionId: "FOOD-score-label",
-    linkIdMatchMode: "strict",
-    skipChart: true,
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const meaningResponse = responses.find((response) => linkIdEquals(response.id, "FOOD-score-label", "strict"));
-      const meaningAnswer =
-        meaningResponse?.answer != null && meaningResponse.answer !== undefined ? meaningResponse.answer : null;
-      return meaningAnswer ? capitalizeFirstLetterSafe(String(meaningAnswer)) : "";
-    },
-  },
-
-  "CIRG-CNICS-FROP-Com": {
-    key: "CIRG-CNICS-FROP-Com",
-    instrumentName: "Falls Risk for Older People in the Community (FROP-Com)",
-    title: "Falls",
-    subtitle: "Past 12 months",
-    questionnaireMatchMode: "fuzzy",
-    displayMeaningNotScore: true,
-    linkIdMatchMode: "strict",
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      let arrMeaning = [];
-      const fallResponse = responses.find((response) => linkIdEquals(response.id, "FROP-Com-0", "strict"));
-      const numFalls = fallResponse?.answer != null && fallResponse.answer !== undefined ? fallResponse.answer : null;
-      if (numFalls !== null) {
-        arrMeaning.push(`Number of falls: ${numFalls.replace(/\bfalls?\b/g, "").trim()}`);
-      }
-      const edVisitResponse = responses.find((response) => linkIdEquals(response.id, "FROP-Com-1", "strict"));
-      const edVisit =
-        edVisitResponse?.answer != null && edVisitResponse.answer !== undefined ? edVisitResponse.answer : null;
-      if (edVisit !== null) {
-        arrMeaning.push(`E/D visit: ${edVisit}`);
-      }
-      return arrMeaning.join("|");
-    },
-    skipMeaningScoreRow: true,
-    skipChart: true,
-  },
-
-  "CIRG-CNICS-HIV-STIGMA": {
-    key: "CIRG-CNICS-HIV-STIGMA",
-    instrumentName: "HIV Stigma",
-    title: "HIV Stigma",
-    alertQuestionId: "HIV-STIGMA-SCORE-CRITICAL",
-    scoringQuestionId: "HIV-STIGMA-SCORE",
-    linkIdMatchMode: "strict",
-    showNumAnsweredWithScore: true,
-    totalAnsweredQuestionId: "HIV-STIGMA-SCORE-NUM-ANSWERED",
-    noteFunction: (questionnaire) => {
-      if (!questionnaire) return "";
-      const matchedItem = questionnaire.item.find((o) => o.linkId === "HIV-STIGMA-SCORE");
-      return matchedItem?.text;
-    },
-    severityBands: [
-      { min: 4, label: "high", meaning: "High Stigma" },
-      { min: 0, label: "low", meaning: "" },
-    ],
-    highSeverityScoreCutoff: 4,
-    chartParams: { ...CHART_CONFIG.default, title: "HIV Stigma Score", xLabel: "" },
-  },
-
-  "CIRG-CNICS-HOUSING": {
-    key: "CIRG-CNICS-HOUSING",
-    instrumentName: "CNICS Housing Measure",
-    title: "Housing",
-    subtitle: "Past month",
-    questionnaireMatchMode: "fuzzy",
-    displayMeaningNotScore: true,
-    linkIdMatchMode: "strict",
-    meaningQuestionId: "HOUSING-1",
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const housingResponse = responses.find((response) => linkIdEquals(response.id, "HOUSING-1", "strict"));
-      const housingAnswer =
-        housingResponse?.answer != null && housingResponse.answer !== undefined ? housingResponse.answer : null;
-      return housingAnswer;
-    },
-    skipMeaningScoreRow: true,
-    skipChart: true,
-  },
-
-  "CIRG-CNICS-IPV4": {
-    key: "CIRG-CNICS-IPV4",
-    instrumentName: "IPV-4",
-    title: "Concern for IPV",
-    subtitle: "Past year",
-    questionnaireMatchMode: "fuzzy",
-    linkIdMatchMode: "strict",
-    highSeverityScoreCutoff: 1,
-    displayMeaningNotScore: true,
-    fallbackScoreMap: {
-      "ipv4-1-0": 1,
-      "ipv4-1-1": 0,
-      "ipv4-2-0": 1,
-      "ipv4-2-1": 0,
-      "ipv4-3-0": 1,
-      "ipv4-3-1": 0,
-      "ipv4-4-0": 1,
-      "ipv4-4-1": 0,
-    },
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      if (!severity || severity === "low") return "";
-      const answerMapping = {
-        "IPV4-1": "Felt trapped",
-        "IPV4-2": "Fearful of harm",
-        "IPV4-3": "Sexual violence",
-        "IPV4-4": "Physical violence",
-      };
-      const answers = responses
-        .filter((response) => {
-          return (
-            (response.answer != null &&
-              response.answer !== undefined &&
-              linkIdEquals(response.linkId, "IPV4-1", "strict")) ||
-            linkIdEquals(response.linkId, "IPV4-2", "strict") ||
-            linkIdEquals(response.linkId, "IPV4-3", "strict") ||
-            linkIdEquals(response.linkId, "IPV4-4", "strict")
-          );
-        })
-        .map((response) => answerMapping[response.linkId]);
-      return answers.join("|");
-    },
-    alertQuestionId: "IPV4-critical",
-    meaningRowLabel: "Summary",
-    skipChart: true,
-  },
-
-  "CIRG-CNICS-MAPSS-SF": {
-    key: "CIRG-CNICS-MAPSS-SF",
-    instrumentName: "Social Support",
-    title: "Social Support",
-    linkIdMatchMode: "strict",
-    skipChart: true,
-    displayMeaningNotScore: true,
-    meaningRowLabel: "Summary",
-    alertQuestionId: "MAPSS-SF-SCORE-CRITICAL",
-    meaningQuestionId: "MAPSS-SF-SCORE-SOCIAL-SUPPORT",
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const meaningResponse = responses.find((response) =>
-        linkIdEquals(response.id, "MAPSS-SF-SCORE-SOCIAL-SUPPORT", "strict"),
-      );
-      const meaningAnswer =
-        meaningResponse?.answer != null && meaningResponse.answer !== undefined ? meaningResponse.answer : null;
-      return meaningAnswer ? capitalizeFirstLetterSafe(String(meaningAnswer)) : "";
-    },
-  },
-
-  "CIRG-CNICS-MINI": {
-    key: "CIRG-CNICS-MINI",
-    instrumentName: "MINI Score",
-    title: "MINI Score",
-    subtitle: "Past 12 months",
-    minimumScore: 0,
-    maximumScore: 7,
-    scoringQuestionId: "MINI-score",
-    meaningQuestionId: "MINI-score-interpretation",
-    excludeQuestionLinkIdPatterns: ["MINI-score-ignoring-skipped"],
-    linkIdMatchMode: "strict",
-    chartParams: {
-      ...CHART_CONFIG.default,
-      title: "MINI Score",
-      minimumYValue: 0,
-      maximumYValue: 7,
-      xLabel: "",
-    },
-  },
-
-  "CIRG-CNICS-SEXUAL-RISK": {
-    key: "CIRG-CNICS-SEXUAL-RISK",
-    instrumentName: "Unprotected Sex",
-    title: "Unprotected Sex",
-    subtitle: "Past 3 months",
-    columns: [
-      {
-        linkId: "SEXUAL-RISK-SCORE-UNPROTECTED",
-        id: "result",
-      },
-    ],
-    valueFormatter: (val) =>
-      String(val).toLowerCase() === "true" ? "Yes" : String(val).toLowerCase() === "false" ? "No" : val,
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const mainResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-UNPROTECTED", "strict"),
-      );
-      const mainResponseAnswer =
-        mainResponse?.answer != null && mainResponse.answer !== undefined ? mainResponse.answer : null;
-      if (!mainResponseAnswer || String(mainResponseAnswer).toLowerCase() !== "yes") return mainResponseAnswer;
-
-      let arrResponses = [];
-      arrResponses.push(mainResponseAnswer);
-
-      // anal sex
-      const analSexResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-UNPROTECTED-ANAL", "strict"),
-      );
-      const analSexAnswer =
-        analSexResponse?.answer != null && analSexResponse.answer !== undefined ? analSexResponse.answer : null;
-      if (String(analSexAnswer).toLowerCase() === "true") arrResponses.push("Anal Sex: Yes");
-      else if (String(analSexAnswer).toLowerCase() === "false") arrResponses.push("Anal Sex: No");
-      else arrResponses.push("Anal Sex: " + (analSexAnswer ? analSexAnswer : "-"));
-
-      // oral sex
-      const oralSexResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-UNPROTECTED-ORAL", "strict"),
-      );
-      const oralSexAnswer =
-        oralSexResponse?.answer != null && oralSexResponse.answer !== undefined ? oralSexResponse.answer : null;
-      if (String(oralSexAnswer).toLowerCase() === "true") arrResponses.push("Oral Sex: Yes");
-      else if (String(oralSexAnswer).toLowerCase() === "false") arrResponses.push("Oral Sex: No");
-      else arrResponses.push("Oral Sex: " + (oralSexAnswer ? oralSexAnswer : "-"));
-
-      // vaginal sex
-      const vaginalSexResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-UNPROTECTED-VAGINAL", "strict"),
-      );
-      const vaginalSexAnswer =
-        vaginalSexResponse?.answer != null && vaginalSexResponse.answer !== undefined
-          ? vaginalSexResponse.answer
-          : null;
-      if (String(vaginalSexAnswer).toLowerCase() === "true") arrResponses.push("Vaginal Sex: Yes");
-      else if (String(vaginalSexAnswer).toLowerCase() === "false") arrResponses.push("Vaginal Sex: No");
-      else arrResponses.push("Vaginal Sex: " + (vaginalSexAnswer ? vaginalSexAnswer : "-"));
-
-      return arrResponses.join("|");
-    },
-    displayMeaningNotScore: true,
-    skipResponses: true,
-    meaningRowLabel: "Unprotected Sex (Past 3 months)",
-    linkIdMatchMode: "strict",
-    skipChart: true,
-  },
-
-  "CIRG-CNICS-Smoking": {
-    key: "CIRG-CNICS-Smoking",
-    instrumentName: "CNICS Smoking",
-    title: "Nicotine Use",
-    questionnaireMatchMode: "fuzzy",
-    displayMeaningNotScore: true,
-    linkIdMatchMode: "strict",
-    excludeQuestionLinkIdPatterns: ["summary"],
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      let arrMeaning = [];
-      const tobaccoUseResponse = responses.find((response) =>
-        linkIdEquals(response.id, "Smoking-Tobacco-Cigs-Summary", "strict"),
-      );
-      const eCigUseResponse = responses.find((response) => linkIdEquals(response.id, "E-Cigarettes-Summary", "strict"));
-      const tobaccoUseAnswer =
-        tobaccoUseResponse?.answer != null && tobaccoUseResponse.answer !== undefined
-          ? tobaccoUseResponse.answer
-          : null;
-      const eCigUseAnswer =
-        eCigUseResponse?.answer != null && eCigUseResponse.answer !== undefined ? eCigUseResponse.answer : null;
-      if (tobaccoUseAnswer) {
-        arrMeaning.push("Tobacco cigarettes: " + tobaccoUseAnswer);
-      }
-      if (eCigUseAnswer) {
-        arrMeaning.push("E-Cigarettes: " + eCigUseAnswer);
-      }
-      return arrMeaning.join("|");
-    },
-    meaningRowLabel: "Summary",
-    skipChart: true,
-  },
-
-  "CIRG-CNICS-Symptoms": {
-    key: "CIRG-CNICS-Symptoms",
-    instrumentName: "CNICS Symptoms Checklist",
-    title: "Current Symptoms",
-    subtitle: "From {date} assessment",
-    questionnaireMatchMode: "fuzzy",
-    linkIdMatchMode: "strict",
-    displayMeaningNotScore: true,
-    meaningRowLabel: "Summary",
-    disableHeaderRowSubtitle: true,
-    columns: [
-      {
-        linkId: "Symptoms-bothers-a-lot",
-        id: "bothersALot",
-      },
-      {
-        linkId: "Symptoms-bothers-some",
-        id: "bothersSome",
-      },
-    ],
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const arrMeaning = [];
-      const bothersALotResponse = responses.find((response) =>
-        linkIdEquals(response.id, "Symptoms-bothers-a-lot", "strict"),
-      );
-      const bothersALotAnswer =
-        bothersALotResponse?.answer != null && bothersALotResponse.answer !== undefined
-          ? bothersALotResponse.answer
-          : null;
-      const bothersSomeResponse = responses.find((response) =>
-        linkIdEquals(response.id, "Symptoms-bothers-some", "strict"),
-      );
-      const bothersSomeAnswer =
-        bothersSomeResponse?.answer != null && bothersSomeResponse.answer !== undefined
-          ? bothersSomeResponse.answer
-          : null;
-      if (bothersALotAnswer) {
-        arrMeaning.push("Bothers a lot: " + bothersALotAnswer);
-      }
-      if (bothersSomeAnswer) {
-        arrMeaning.push("Bothers some: " + bothersSomeAnswer);
-      }
-      return arrMeaning.join("|");
-    },
-    skipChart: true,
-  },
-
+  "CIRG-ADL-IADL": CIRG_ADL_IADL,
+  "CIRG-BEHAV5": CIRG_BEHAV5,
+  "CIRG-C-IDAS": CIRG_C_IDAS,
+  "CIRG-CNICS-ARV": CIRG_CNICS_ARV,
+  "CIRG-CNICS-ASSIST": CIRG_CNICS_ASSIST,
+  "CIRG-CNICS-ASSIST-OD": CIRG_CNICS_ASSIS_OD,
+  "CIRG-CNICS-ASSIST-Polysub": CIRG_CNICS_ASSIST_POLYSUB,
+  "CIRG-CNICS-AUDIT": CIRG_CNICS_AUDIT,
+  "CIRG-CNICS-EUROQOL": CIRG_CNICS_EUROQOL,
+  "CIRG-CNICS-EUROQOL-ANXIETY-DEPRESSION": CIRG_CNICS_EUROQOL_ANXIETY_DEPRESSION,
+  "CIRG-CNICS-EUROQOL-EUROQOL-5": CIRG_CNICS_EUROQOL_EUROQOL_5,
+  "CIRG-CNICS-EUROQOL-PAIN-DISCOMFORT": CIRG_CNICS_EUROQOL_PAIN_DISCOMFORT,
+  "CIRG-CNICS-EUROQOL-SELF-CARE": CIRG_CNICS_EUROQOL_SELF_CARE,
+  "CIRG-CNICS-EUROQOL-USUAL-ACTIVITIES": CIRG_CNICS_EUROQOL_USUAL_ACTIVITIES,
+  "CIRG-CNICS-EXCHANGE-SEX": CIRG_CNICS_EXCHANGE_SEX,
+  "CIRG-CNICS-FINANCIAL": CIRG_CNICS_FINANCIAL,
+  "CIRG-CNICS-FOOD": CIRG_CNICS_FOOD,
+  "CIRG-CNICS-FROP-Com": CIRG_CNICS_FROP_COM,
+  "CIRG-CNICS-HIV-STIGMA": CIRG_CNICS_HIV_STIGMA,
+  "CIRG-CNICS-HOUSING": CIRG_CNICS_HOUSING,
+  "CIRG-CNICS-IPV4": CIRG_CNICS_IPV4,
+  "CIRG-CNICS-MAPSS-SF": CIRG_CNICS_MAPSS_SF,
+  "CIRG-CNICS-MINI": CIRG_CNICS_MINI,
+  "CIRG-CNICS-SEXUAL-RISK": CIRG_CNICS_SEXUAL_RISK,
+  "CIRG-CNICS-Smoking": CIRG_CNICS_SMOKING,
+  "CIRG-CNICS-Symptoms": CIRG_CNICS_SYMPTOMS,
   "CIRG-CONCURRENT-IDU": {
-    key: "CIRG-CONCURRENT-IDU",
     instrumentName: "Concurrent IDU",
     title: "Concurrent IDU",
   },
-
-  "CIRG-CP-ECOG": {
-    key: "CIRG-CP-ECOG",
-    questionnaireId: "CIRG-CP-ECOG",
-    questionnaireName: "cp-ecog",
-    instrumentName: "CP-ECOG",
-    title: "CP-ECOG",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/cp-ecog",
-    scoringQuestionId: "cp-ecog-total-score",
-    maximumScore: 48,
-    questionnaireMatchMode: "fuzzy",
-    questionLinkIds: [
-      "/89286-9/89146-5",
-      "/89286-9/89149-9",
-      "/89287-7/89172-1",
-      "/89287-7/89138-2",
-      "/89288-5/89154-9",
-      "/89288-5/89165-5",
-      "/89289-3/89143-2",
-      "/89289-3/89140-8",
-      "/89290-1/89158-0",
-      "/89290-1/89173-9",
-      "/89285-1/89141-6",
-      "/89285-1/89171-3",
-    ],
-    chartParams: { ...CHART_CONFIG.default, minimumYValue: 0, maximumYValue: 48, xLabel: "" },
-  },
-
-  "CIRG-ECOG12": {
-    key: "CIRG-ECOG12",
-    questionnaireId: "CIRG-ECOG12",
-    questionnaireName: "ecog12",
-    instrumentName: "ECOG-12",
-    title: "ECOG-12",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/ecog12",
-    scoringQuestionId: "ecog12-total-score",
-    maximumScore: 48,
-    questionnaireMatchMode: "fuzzy",
-    questionLinkIds: [
-      "/89286-9/89146-5",
-      "/89286-9/89149-9",
-      "/89287-7/89172-1",
-      "/89287-7/89138-2",
-      "/89288-5/89154-9",
-      "/89288-5/89165-5",
-      "/89289-3/89143-2",
-      "/89289-3/89140-8",
-      "/89290-1/89158-0",
-      "/89290-1/89173-9",
-      "/89285-1/89141-6",
-      "/89285-1/89171-3",
-    ],
-    chartParams: { ...CHART_CONFIG.default, minimumYValue: 0, maximumYValue: 48, xLabel: "" },
-  },
-
+  "CIRG-CP-ECOG": CIRG_CP_ECOG,
+  "CIRG-ECOG12": CIRG_ECOG12,
   "CIRG-FENTANYL-STRIP-ACCESS": {
-    key: "CIRG-FENTANYL-STRIP-ACCESS",
     instrumentName: "Fentanyl Test Strip Access",
     title: "Fentanyl Test Strip Access",
   },
-
-  "CIRG-GAD7": {
-    key: "CIRG-GAD7",
-    questionnaireId: "CIRG-GAD7",
-    questionnaireName: "gad7",
-    instrumentName: "GAD-7",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/gad7",
-    scoringQuestionId: "/70274-6",
-    maximumScore: 21,
-    questionLinkIds: ["/69725-0", "/68509-9", "/69733-4", "/69734-2", "/69735-9", "/69689-8", "/69736-7"],
-    questionnaireMatchMode: "fuzzy",
-    severityBands: [
-      { min: 15, label: "high", meaning: "severe anxiety" },
-      { min: 10, label: "moderate", meaning: "moderate anxiety" },
-      { min: 5, label: "mild", meaning: "mild anxiety" },
-      { min: 0, label: "low", meaning: "minimal anxiety" },
-    ],
-    // optional (defaults to top band min = 15 anyway)
-    highSeverityScoreCutoff: 15,
-    mediumSeverityScoreCutoff: 10,
-    chartParams: { ...CHART_CONFIG.default, minimumYValue: 0, maximumYValue: 21, xLabel: "", dotColor: null },
-  },
-
-  "CIRG-GDS": {
-    key: "CIRG-GDS",
-    questionnaireId: "CIRG-GDS",
-    questionnaireName: "gds",
-    instrumentName: "GDS",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/gds",
-    scoringQuestionId: "/48545-8",
-    maximumScore: 15,
-    questionnaireMatchMode: "fuzzy",
-    questionLinkIds: [
-      "/48512-8",
-      "/48513-6",
-      "/48514-4",
-      "/48515-1",
-      "/48518-5",
-      "/48519-3",
-      "/48520-1",
-      "/48521-9",
-      "/48523-5",
-      "/48525-0",
-      "/48526-8",
-      "/48528-4",
-      "/48532-6",
-      "/48533-4",
-      "/48534-2",
-    ],
-    chartParams: { ...CHART_CONFIG.default, minimumYValue: 0, maximumYValue: 15, xLabel: "" },
-  },
-
-  //TODO, implement those
-
+  "CIRG-GAD7": CIRG_GAD7,
+  "CIRG-GDS": CIRG_GDS,
   "CIRG-IDU": {
-    key: "CIRG-IDU",
     instrumentName: "IDU",
     title: "IDU",
   },
-
-  "CIRG-LAST-MISSED-DOSE": {
-    key: "CIRG-LAST-MISSED-DOSE",
-    title: "Last Missed Dose",
-    subtitle: "Past 4 weeks",
-    instrumentName: "Last Missed Dose",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-ARV"], // one or many hosts
-      linkId: "ARV-last-missed", // the single item to keep
-    },
-    columns: [
-      {
-        linkId: "ARV-last-missed",
-        id: "result",
-      },
-    ],
-    skipMeaningScoreRow: true,
-    skipChart: true,
-  },
-
-  "CIRG-MINICOG": {
-    key: "CIRG-MINICOG",
-    questionnaireId: "CIRG-MINICOG",
-    questionnaireName: "MINICOG",
-    instrumentName: "MINI-COG",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/minicog",
-    recallLinkIds: ["/recall-1", "/recall-2", "/recall-3"],
-    clockLinkId: "/clock",
-    // What counts as "correct" for recall (pick your reality):
-    recallCorrectCodes: ["correct", "present", "Y"], // if coded
-    recallCorrectStrings: ["correct", "present", "yes"], // if strings
-    // Clock scoring: map whichever value you use to 0 or 2
-    clockScoreMap: {
-      normal: 2,
-      abnormal: 0,
-      2: 2,
-      0: 0,
-    },
-    // Total severity bands for Mini-Cog
-    severityBands: [
-      { min: 3, label: "low", meaning: "unlikely impairment" },
-      { min: 0, label: "high", meaning: "possible impairment" }, // catch-all for 0–2
-    ],
-    highSeverityScoreCutoff: 2, // flag when total ≤ 2 (common rule)
-    comparisonToAlert: "lower",
-    questionnaireMatchMode: "fuzzy",
-    chartParams: { ...CHART_CONFIG.default, dotColor: null },
-  },
-
+  "CIRG-LAST-MISSED-DOSE": CIRG_CNICS_ARV_MISSED_DOSE,
+  "CIRG-MINICOG": CIRG_MINICOG,
   "CIRG-NALOXONE-ACCESS": {
-    key: "CIRG-NALOXONE-ACCESS",
     instrumentName: "Naloxone Access",
     title: "Naloxone Access",
   },
-
-  "CIRG-PARTNER-CONTEXT": {
-    key: "CIRG-PARTNER-CONTEXT",
-    instrumentName: "Sexual Partner Context",
-    title: "Sexual Partner Context",
-    subtitle: "Past 3 months",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-SEXUAL-RISK"], // one or many hosts
-      linkIds: [
-        "SEXUAL-RISK-SCORE-PARTNERS-GENDERS",
-        "SEXUAL-RISK-SCORE-PARTNERS-HIV-NEG-PREP",
-        "SEXUAL-RISK-SCORE-PARTNERS-HIV-UNKNOWN",
-      ],
-    },
-    columns: [
-      {
-        linkId: "SEXUAL-RISK-SCORE-PARTNERS-GENDERS",
-        id: "result",
-      },
-      {
-        linkId: "SEXUAL-RISK-SCORE-PARTNERS-HIV-NEG-PREP",
-        id: "result",
-      },
-      {
-        linkId: "SEXUAL-RISK-SCORE-PARTNERS-HIV-UNKNOWN",
-        id: "result",
-      },
-    ],
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      let arrResponses = [];
-
-      // gender
-      const genderSexResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-PARTNERS-GENDERS", "strict"),
-      );
-      const genderSexAnswer =
-        genderSexResponse?.answer != null && genderSexResponse.answer !== undefined ? genderSexResponse.answer : null;
-      if (genderSexAnswer && String(genderSexAnswer).toLowerCase() !== "tbd") arrResponses.push(genderSexAnswer);
-
-      // PREP
-      const prepSexResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-PARTNERS-HIV-NEG-PREP", "strict"),
-      );
-      const prepSexAnswer =
-        prepSexResponse?.answer != null && prepSexResponse.answer !== undefined ? prepSexResponse.answer : null;
-      if (prepSexAnswer && String(prepSexAnswer).toLowerCase() !== "tbd") arrResponses.push(prepSexAnswer);
-
-      // HIV unknown
-      const hivUnknownResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-PARTNERS-HIV-UNKNOWN", "strict"),
-      );
-      const hivUnknownAnswer =
-        hivUnknownResponse?.answer != null && hivUnknownResponse.answer !== undefined
-          ? hivUnknownResponse.answer
-          : null;
-      if (hivUnknownAnswer && String(hivUnknownAnswer).toLowerCase() !== "tbd") arrResponses.push(hivUnknownAnswer);
-
-      return arrResponses.join("|");
-    },
-    displayMeaningNotScore: true,
-    skipResponses: true,
-    meaningRowLabel: "Sexual Partner Context (Past 3 months)",
-    linkIdMatchMode: "strict",
-    skipChart: true,
-  },
-
-  "CIRG-PC-PTSD-5": {
-    key: "CIRG-PC-PTSD-5",
-    questionnaireId: "CIRG-PC-PTSD-5",
-    questionnaireName: "CIRG-PC-PTSD-5",
-    instrumentName: "The Primary Care PTSD Screen for DSM-5 [PC-PTSD-5]",
-    title: "PTSD Symptoms",
-    subtitle: "Past month",
-    //  skipMeaningScoreRow: true,
-    skipChart: true,
-    meaningQuestionId: "PC-PTSD-5-SCORE-SYMPTOMS",
-    displayMeaningNotScore: true,
-    linkIdMatchMode: "strict",
-    excludeQuestionLinkIdPatterns: ["102017-1"],
-    questionnaireMatchMode: "fuzzy",
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const meaningResponse = responses.find((response) =>
-        linkIdEquals(response.id, "PC-PTSD-5-SCORE-SYMPTOMS", "strict"),
-      );
-      const meaningAnswer =
-        meaningResponse?.answer != null && meaningResponse.answer !== undefined ? meaningResponse.answer : null;
-      return meaningAnswer?.split(",").join("|");
-    },
-    meaningRowLabel: "Summary (Symptoms endorsed in the past month)",
-    disableHeaderRowSubtitle: true,
-  },
-
-  "CIRG-PHQ9": {
-    key: "CIRG-PHQ9",
-    questionnaireId: "CIRG-PHQ9",
-    questionnaireName: "phq9",
-    instrumentName: "Patient Health Questionnaire-9 (PHQ-9)",
-    title: "PHQ-9",
-    subtitle: "Last two weeks",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/phq9",
-    scoringQuestionId: "/44261-6",
-    note: PHQ9_ADMIN_NOTE,
-    subScoringQuestions: [
-      {
-        key: "PHQ-2",
-        linkId: "/55758-7",
-      },
-    ],
-    maximumScore: 27,
-    linkIdMatchMode: "fuzzy",
-    questionLinkIds: [
-      "/55758-7",
-      "/44250-9",
-      "/44255-8",
-      "/44259-0",
-      "/44254-1",
-      "/44251-7",
-      "/44258-2",
-      "/44252-5",
-      "/44253-3",
-      "/44260-8",
-      //"/69722-7",
-      //"/44261-6",
-      //"/55758-7",
-    ],
-    itemTextByLinkId: {
-      "/55758-7": "PHQ-2 total score",
-      "/44261-6": "PHQ-9 total score",
-    },
-    questionnaireMatchMode: "fuzzy",
-    highSeverityScoreCutoff: 20,
-    mediumSeverityScoreCutoff: 10,
-    severityBands: [
-      { min: 20, label: "high", meaning: "Severe depression" },
-      { min: 15, label: "moderately high", meaning: "Moderately severe depression" },
-      { min: 10, label: "moderate", meaning: "Moderate depression" },
-      { min: 5, label: "mild", meaning: "Mild depression" },
-      { min: 0, label: "low", meaning: "Minimal depression" },
-    ],
-    showNumAnsweredWithScore: true,
-    chartParams: {
-      ...CHART_CONFIG.default,
-      title: "PHQ-9 Score",
-      minimumYValue: 0,
-      maximumYValue: 27,
-      xLabel: "",
-      connectNulls: true,
-      dotColor: null,
-      splitBySource: true,
-    },
-  },
-
-  "CIRG-SEXUAL-PARTNERS": {
-    key: "CIRG-SEXUAL-PARTNERS",
-    instrumentName: "Number of Sexual Partners",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-SEXUAL-RISK"], // one or many hosts
-      linkId: "SEXUAL-RISK-SCORE-NUM-PARTNERS", // the single item to keep
-    },
-    columns: [
-      {
-        linkId: "SEXUAL-RISK-SCORE-NUM-PARTNERS",
-        id: "result",
-      },
-    ],
-    title: "# of Sex Partners",
-    subtitle: "Past 3 months",
-    skipMeaningScoreRow: true,
-    skipChart: true,
-  },
-
+  "CIRG-PARTNER-CONTEXT": CIRG_SEXUAL_PARTNER_CONTEXT,
+  "CIRG-PC-PTSD-5": CIRG_PC_PTSD_5,
+  "CIRG-PHQ9": CIRG_PHQ9,
+  "CIRG-SEXUAL-PARTNERS": CIRG_SEXUAL_PARTNERS,
   "CIRG-SHORTNESS-OF-BREATH": {
     title: "Shortness of breath",
   },
-
-  "CIRG-SI": {
-    key: "CIRG-SI",
-    instrumentName: "Suicide Ideation",
-    title: "Suicide Ideation",
-    subtitle: "Last two weeks",
-    scoringQuestionId: PHQ9_SI_QUESTION_LINK_ID,
-    fallbackScoreMap: PHQ9_SI_ANSWER_SCORE_MAPPINGS,
-    highSeverityScoreCutoff: 3,
-    mediumSeverityScoreCutoff: 2,
-    comparisonToAlert: "higher",
-    severityBands: [
-      { min: 3, label: "high", meaning: "Nearly every day" },
-      { min: 2, label: "moderate", meaning: "More than half the days" },
-      { min: 1, label: "mild", meaning: "Several days" },
-      { min: 0, label: "low", meaning: "Not at all" },
-    ],
-    minimumScore: 0,
-    maximumScore: 3,
-    deriveFrom: {
-      hostIds: ["CIRG-PHQ9"], // one or many hosts
-      linkId: PHQ9_SI_QUESTION_LINK_ID, // the single item to keep
-    },
-    chartParams: {
-      ...CHART_CONFIG.default,
-      title: "Suicide Ideation",
-      minimumYValue: 0,
-      maximumYValue: 3,
-      xLabel: "",
-      yLabel: "value",
-      type: "barchart",
-      dotColor: null,
-    },
-  },
-
-  "CIRG-SLUMS": {
-    key: "CIRG-SLUMS",
-    questionnaireId: "CIRG-SLUMS",
-    questionnaireName: "slums",
-    instrumentName: "SLUMS",
-    questionnaireUrl: "http://www.cdc.gov/ncbddd/fasd/slums",
-    scoringQuestionId: "/71492-3", // total score item
-    questionLinkIds: ["/71492-3"],
-    maximumScore: 30,
-    comparisonToAlert: "lower",
-    questionnaireMatchMode: "fuzzy",
-    // No questionLinkIds needed—SLUMS uses a single total-score field
-    chartParams: { ...CHART_CONFIG.default, title: "SLUMS", minimumYValue: 0, maximumYValue: 30, xLabel: "" },
-  },
-
-  "CIRG-SRS": {
-    key: "CIRG-SRS",
-    title: "Self Rating Scale (SRS)",
-    subtitle: "Past 4 weeks",
-    instrumentName: "Self Rating Scale (SRS)",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-ARV"], // one or many hosts
-      linkId: "ARV-SRS", // the single item to keep
-    },
-    columns: [
-      {
-        linkId: "SARV-SRS",
-        id: "result",
-      },
-    ],
-    skipMeaningScoreRow: true,
-    skipChart: true,
-  },
-
-  "CIRG-STI": {
-    key: "CIRG-STI",
-    instrumentName: "STI",
-    title: "Concern for STI Exposure",
-    subtitle: "Past 3 months",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-SEXUAL-RISK"], // one or many hosts
-      linkId: "SEXUAL-RISK-SCORE-STI-EXPOSURE", // the single item to keep
-    },
-    columns: [
-      {
-        linkId: "SEXUAL-RISK-SCORE-STI-EXPOSURE",
-        id: "result",
-      },
-    ],
-    valueFormatter: (val) =>
-      String(val).toLowerCase() === "true" ? "Yes" : String(val).toLowerCase() === "false" ? "No" : "",
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const meaningResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-STI-EXPOSURE", "strict"),
-      );
-      const meaningAnswer =
-        meaningResponse?.answer != null && meaningResponse.answer !== undefined ? meaningResponse.answer : null;
-      if (String(meaningAnswer).toLowerCase() === "true") return "Yes";
-      else if (String(meaningAnswer).toLowerCase() === "false") return "No";
-      return meaningAnswer;
-    },
-    displayMeaningNotScore: true,
-    skipResponses: true,
-    meaningRowLabel: "Concern for STI Exposure (Last 3 months)",
-    linkIdMatchMode: "strict",
-    alertQuestionId: "SEXUAL-RISK-SCORE-STI-EXPOSURE",
-    skipChart: true,
-  },
-
-  "CIRG-UNPROTECTED-ANAL-SEX": {
-    key: "CIRG-UNPRTECTED-ANAL-SEX",
-    instrumentName: "Unprotected Sex",
-    title: "Unprotected Anal Sex",
-    subtitle: "Past 3 months",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-SEXUAL-RISK"], // one or many hosts
-      linkId: "SEXUAL-RISK-SCORE-UNPROTECTED-ANAL", // the single item to keep
-    },
-    columns: [
-      {
-        linkId: "SEXUAL-RISK-SCORE-UNPROTECTED-ANAL",
-        id: "result",
-      },
-    ],
-    valueFormatter: (val) =>
-      String(val).toLowerCase() === "true" ? "Yes" : String(val).toLowerCase() === "false" ? "No" : val,
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const meaningResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-UNPROTECTED-ANAL", "strict"),
-      );
-      const meaningAnswer =
-        meaningResponse?.answer != null && meaningResponse.answer !== undefined ? meaningResponse.answer : null;
-      if (String(meaningAnswer).toLowerCase() === "true") return "Yes";
-      else if (String(meaningAnswer).toLowerCase() === "false") return "No";
-      return meaningAnswer;
-    },
-    displayMeaningNotScore: true,
-    skipResponses: true,
-    meaningRowLabel: "Unprotected Anal Sex",
-    linkIdMatchMode: "strict",
-    skipChart: true,
-  },
-
-  "CIRG-UNPROTECTED-ORAL-SEX": {
-    key: "CIRG-UNPROTECTED-ORAL-SEX",
-    instrumentName: "Unprotected Sex",
-    title: "Unprotected Oral Sex",
-    subtitle: "Past 3 months",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-SEXUAL-RISK"], // one or many hosts
-      linkId: "SEXUAL-RISK-SCORE-UNPROTECTED-ORAL", // the single item to keep
-    },
-    columns: [
-      {
-        linkId: "SEXUAL-RISK-SCORE-UNPROTECTED-ORAL",
-        id: "result",
-      },
-    ],
-    valueFormatter: (val) =>
-      String(val).toLowerCase() === "true" ? "Yes" : String(val).toLowerCase() === "false" ? "No" : val,
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const meaningResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-UNPROTECTED-ORAL", "strict"),
-      );
-      const meaningAnswer =
-        meaningResponse?.answer != null && meaningResponse.answer !== undefined ? meaningResponse.answer : null;
-      if (String(meaningAnswer).toLowerCase() === "true") return "Yes";
-      else if (String(meaningAnswer).toLowerCase() === "false") return "No";
-      return meaningAnswer;
-    },
-    displayMeaningNotScore: true,
-    skipResponses: true,
-    meaningRowLabel: "Unprotected Oral Sex",
-    linkIdMatchMode: "strict",
-    skipChart: true,
-  },
-
-  "CIRG-UNPROTECTED-VAGINAL-SEX": {
-    key: "CIRG-UNPROTECTED-VAGINAL-SEX",
-    instrumentName: "Unprotected Sex",
-    title: "Unprotected Vaginal Sex",
-    subtitle: "Past 3 months",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-SEXUAL-RISK"], // one or many hosts
-      linkId: "SEXUAL-RISK-SCORE-UNPROTECTED-VAGINAL", // the single item to keep
-    },
-    columns: [
-      {
-        linkId: "SEXUAL-RISK-SCORE-UNPROTECTED-VAGINAL",
-        id: "result",
-      },
-    ],
-    valueFormatter: (val) =>
-      String(val).toLowerCase() === "true" ? "Yes" : String(val).toLowerCase() === "false" ? "No" : val,
-    fallbackMeaningFunc: function (severity, responses) {
-      if (isEmptyArray(responses)) return "";
-      const meaningResponse = responses.find((response) =>
-        linkIdEquals(response.id, "SEXUAL-RISK-SCORE-UNPROTECTED-VAGINAL", "strict"),
-      );
-      const meaningAnswer =
-        meaningResponse?.answer != null && meaningResponse.answer !== undefined ? meaningResponse.answer : null;
-      if (String(meaningAnswer).toLowerCase() === "true") return "Yes";
-      else if (String(meaningAnswer).toLowerCase() === "false") return "No";
-      return meaningAnswer;
-    },
-    displayMeaningNotScore: true,
-    skipResponses: true,
-    meaningRowLabel: "Unprotected Vaginal Sex",
-    linkIdMatchMode: "strict",
-    skipChart: true,
-  },
-
-  "CIRG-VAS": {
-    key: "CIRG-VAS",
-    title: "Percent ART taken",
-    subtitle: "Past 4 weeks",
-    instrumentName: "VAS",
-    scoringQuestionId: "ARV-VAS",
-    deriveFrom: {
-      hostIds: ["CIRG-CNICS-ARV"], // one or many hosts
-      linkId: "ARV-VAS", // the single item to keep
-    },
-    columns: [
-      {
-        linkId: "ARV-VAS",
-        id: "result",
-      },
-    ],
-    valueFormatter: (value) => (!isNil(value) ? `${value} ${String(value).includes("%") ? "" : "%"}` : value),
-    chartParams: {
-      ...CHART_CONFIG.default,
-      title: "Percent ART Taken",
-      minimumYValue: 0,
-      maximumYValue: 100,
-      xLabel: "",
-      yLabel: "value",
-      tooltipValueFormatter: (value) => (!isNil(value) ? `${value} ${String(value).includes("%") ? "" : "%"}` : value),
-      type: "barchart",
-    },
-    skipResponses: true,
-    meaningRowLabel: "Visual Analog Scale % (Past 4 weeks)",
-  },
+  "CIRG-SI": CIRG_SI,
+  "CIRG-SLUMS": CIRG_SLUMS,
+  "CIRG-SRS": CIRG_CNICS_ARV_SRS,
+  "CIRG-STI": CIRG_STI,
+  "CIRG-UNPROTECTED-ANAL-SEX": CIRG_UNPROTECTED_ANAL_SEX,
+  "CIRG-UNPROTECTED-ORAL-SEX": CIRG_UNPROTECTED_ORAL_SEX,
+  "CIRG-UNPROTECTED-VAGINAL-SEX": CIRG_UNPROTECTED_VAGINAL_SEX,
+  "CIRG-VAS": CIRG_CNICS_ARV_VAS,
 };
-
-console.log("sorted ", Object.fromEntries(
-  Object.entries(questionnaireConfigsRaw).sort(([a], [b]) => a.localeCompare(b))
-))
 
 export const getConfigForQuestionnaire = (id) => {
   return (
